requirements.txt
appdirs==1.4.3
backports.functools-lru-cache==1.6.1
chainer==1.17.0
cloudpickle==1.3.0
configparser==4.0.2
contextlib2==0.6.0.post1
cycler==0.10.0
decorator==4.4.2
dill==0.3.3
distlib==0.3.0
filelock==3.0.12
importlib-metadata==1.6.0
importlib-resources==1.4.0
kiwisolver==1.1.0
matplotlib==2.2.5
networkx==2.2
nose==1.3.7
numpy==1.16.6
pathlib2==2.3.5
Pillow==6.2.2
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.0.3
scandir==1.10.0
scikit-image==0.14.5
scikit-learn==0.20.4
scipy==1.2.3
singledispatch==3.4.0.3
six==1.14.0
subprocess32==3.5.4
typing==3.7.4.1
virtualenv==20.0.18
zipp==1.2.0
psutil==5.8.0
tensorflow-gpu==1.15.0
torch==1.4.0
future==0.18.2
torchvision==0.5.0
pandas==0.24.2
